Move GroupMLP layer to layers
